fix(docs): keep metadata helper import-safe
